Select a derived entity
