+++ context/app/static/js/components/detailPage/related-entities/RelatedEntitiesTable/index.ts
@@ -1,0 +1,3 @@
+import RelatedEntitiesTable from './RelatedEntitiesTable';
+
+export default RelatedEntitiesTable;
